Fix admin resume tab type

+++ lib/admin_resume_session.dart
@@ -61,7 +61,7 @@ class AdminResumeSession {
         await clear();
         return null;
       }
-      final tab = (prefs.getInt(_tabKey) ?? 0).clamp(0, 7);
+      final tab = (prefs.getInt(_tabKey) ?? 0).clamp(0, 7).toInt();
       return AdminResumeSnapshot(secret: secret, tab: tab);
     } catch (_) {
       return null;
